Moved Find and ValidateRefForInput into AssetFactory namespace.
